+++ src/muncher/races/races.js
@@ -3,6 +3,7 @@ import logger from "../../logger.js";
 import DDBRace from "../../parser/race/DDBRace.js";
 import DDBItemImporter from "../../lib/DDBItemImporter.js";
 import DDBRaceTrait from "../../parser/race/DDBRaceTrait.js";
+import { DDBCompendiumFolders } from "../../lib/DDBCompendiumFolders.js";
 
 const NO_TRAITS = [
   "Speed",
@@ -14,6 +15,13 @@ const NO_TRAITS = [
   "Extra Language",
   "Age",
   "Alignment",
+  "Creature Type",
+  "Darkvision",
+  "Keen Senses",
+  "Natural Athlete",
+  "Necrotic Resistance",
+  "Skills",
+  "Skill Versatility",
 ];
 
 export async function getRaces(data) {
@@ -24,29 +32,38 @@ export async function getRaces(data) {
   let races = [];
   let racialFeatures = [];
 
+  const traitCompendiumFolders = new DDBCompendiumFolders("traits");
+  await traitCompendiumFolders.loadCompendium("traits");
+
   const excludeLegacy = game.settings.get("ddb-importer", "munching-policy-exclude-legacy");
-  data
-    .filter((race) => !excludeLegacy || (excludeLegacy && !race.isLegacy))
-    .forEach((race) => {
-      logger.debug(`${race.fullName} features parsing started...`);
-      race.racialTraits.forEach((trait) => {
-        logger.debug(`${trait.definition.name} trait starting...`);
-        if (!trait.definition.hideInSheet && !NO_TRAITS.includes(trait.definition.name)) {
-          const ddbTrait = new DDBRaceTrait(trait.definition, race.fullName, { isLegac: race.isLegacyy });
-          racialFeatures.push(ddbTrait.data);
-          results.push({ race: race.fullName, trait: trait.definition.name });
-        }
-      });
-    });
+  const filteredRaces = data.filter((race) => !excludeLegacy || (excludeLegacy && !race.isLegacy));
+
+  for (const race of filteredRaces) {
+    logger.debug(`${race.fullName} features parsing started...`);
+    const groupName = DDBRace.getGroupName(race.groupIds, race.baseRaceName);
+    // await traitCompendiumFolders.getRacialBaseFolder("trait", groupName);
+    for (const trait of race.racialTraits) {
+      logger.debug(`${trait.definition.name} trait starting...`);
+      if (!trait.definition.hideInSheet && !NO_TRAITS.includes(trait.definition.name)) {
+        const ddbTrait = new DDBRaceTrait(trait.definition, race);
+        racialFeatures.push(ddbTrait.data);
+        results.push({ race: race.fullName, trait: trait.definition.name });
+        await traitCompendiumFolders.createSubTraitFolders(groupName, race.fullName);
+      }
+    }
+  }
+
+  const raceCompendiumFolders = new DDBCompendiumFolders("races");
+  await raceCompendiumFolders.loadCompendium("races");
 
   const traitHelper = await DDBItemImporter.buildHandler("traits", racialFeatures, updateBool, { chrisPremades: true, matchFlags: ["entityRaceId"] });
   const compendiumRacialTraits = await DDBRace.getRacialTraitsLookup(traitHelper.documents);
-  const filteredRaces = data.filter((race) => !excludeLegacy || (excludeLegacy && !race.isLegacy));
 
   for (const race of filteredRaces) {
     logger.debug(`${race.fullName} race parsing started...`);
     const ddbRace = new DDBRace(null, race, compendiumRacialTraits);
     await ddbRace.build();
+    await raceCompendiumFolders.getRacialBaseFolder("race", ddbRace.groupName);
     races.push(ddbRace.data);
   }