+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Effect: Heroism updated to MidiQoL Overtime effect
 - DDB Importer v4.0.0 no longer supports Foundry 10, or 5e versions earlier than v3.0.0.
 - Item Containers/Collections module is no longer supported in favour of native Foundry containers.
-
+- Race and Racial Traits now import into compendium folders.
 
 Note: There are still a number of improvements to come over the coming week(s) to the new rule referencing techniques in descriptions and journals.